moved some stuff around and reader check
