+++ include/pangolin/utils/file_utils.h
@@ -34,6 +34,14 @@
 #include <vector>
 #include <algorithm>
 
+#ifndef _WIN_
+#include <sys/stat.h>
+#include <sys/signal.h>
+#include <stdio.h>
+#include <fcntl.h>
+#include <errno.h>
+#endif // _WIN_
+
 namespace pangolin
 {
 
@@ -114,6 +122,31 @@ inline std::string ToLowerCopy( const std::string& str )
     out.resize(str.size());
     std::transform(str.begin(), str.end(), out.begin(), ::tolower);
     return out;
+}
+
+inline bool IsPipe(const std::string& file)
+{
+#ifdef _WIN_
+    return false;
+#else
+    struct stat st;
+    stat(file.c_str(), &st);
+    return ((st.st_mode & S_IFMT) == S_IFIFO);
+#endif // _WIN_
+}
+
+inline bool PipeHasReader(const std::string& file)
+{
+#ifdef _WIN_
+    return false;
+#else
+    int ret = ::open(file.c_str(), O_WRONLY | O_NONBLOCK);
+    if(ret == -1)
+    {
+        return errno != ENXIO;
+    }
+    return true;
+#endif // _WIN_
 }
 
 }
