+++ include/pangolin/video/video_output.h
@@ -46,26 +46,8 @@
 
 #include <pangolin/video/video.h>
 
-#ifdef _WIN_
-#else
-#include <sys/stat.h>
-#include <stdio.h>
-#endif // _WIN_
-
-
 namespace pangolin
 {
-
-inline bool IsPipe(const std::string& file)
-{
-#ifdef _WIN_
-    return false;
-#else
-    struct stat st;
-    stat(file.c_str(), &st);
-    return ((st.st_mode & S_IFMT) == S_IFIFO);
-#endif // _WIN_
-}
 
 //! Interface to video recording destinations
 struct PANGOLIN_EXPORT VideoOutputInterface
